+++ windows/ingredientsWindow/delete_iWindow.java
@@ -39,7 +39,7 @@ public class delete_iWindow extends abstractEdit_CheckWindow {
             String id = Integer.toString(temp.getId());
             String prov_id = Integer.toString(temp.getProviderID());
             String price = Float.toString(temp.getPrice());
-            String amount = Integer.toString(temp.getAmount());
+            String amount = Float.toString(temp.getAmount());
             String in_inventory = "No";
             if (temp.getInInventory())
                 in_inventory = "Yes";
