+++ windows/ingredientsWindow/edit_iWindow.java
@@ -48,7 +48,7 @@ public class edit_iWindow extends abstractEdit_CheckWindow {
             String id = Integer.toString(temp.getId());
             String prov_id = Integer.toString(temp.getProviderID());
             String price = Float.toString(temp.getPrice());
-            String amount = Integer.toString(temp.getAmount());
+            String amount = Float.toString(temp.getAmount());
             String in_inventory = "No";
             if (temp.getInInventory())
                 in_inventory = "Yes";
